added a new ts.iteration-based openPMD track converter
